Add real opcodes to lexer and parser
Lots of work in this commit. Add an Opcode table to the module level
to be used by both the lexer and the parser. Write a get_opcode_byte
function to dynamically calculate the opcode byte the way that
uxnasm.c does it. Add a value part to Tokens. Fixup _add_token() debug
printing to show the value. Re-enable debug printing for LEXER stuff.
Add _handle_opcode to parser. Use raw values for JCI and JMI like
unxasm.c does because it doesn't support writing them out anyway,
afaict.